Split test file for a more synoptic project structure
